Fix subfile compilation of chpaters with subfiles package.
Move loading of tikz sooner to which fixes compilation with lualatex of
chapter subfiles.

Remove full definition of Minion3 definition (workaround for bug in
microtype after fontspec update. Now relies on external fontspec
definition but this is not core to the package as it requires and
external font to be installed.

more code clean up.

Signed-off-by: Gregory J. L. Tourte <greg.tourte@gmail.com>